% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atm_var_station.R
\name{atm_var_station}
\alias{atm_var_station}
\title{Get data for the atmospheric variable of your choice and the altitudes of your choice for the closest and the furthest four surrounding points of your station of choice}
\usage{
atm_var_station(
  sta_name,
  data_dir,
  altitudes,
  variable,
  folder,
  itx3d,
  stations,
  days
)
}
\arguments{
\item{sta_name}{Give the station name of your choice, it needs to be in between "" and match the names in the stations data frame that you also need to input in the function}

\item{data_dir}{Give your data folder, don't give the whole path, just the name of the folder inside the working directory}

\item{altitudes}{Give a vector or a value of altitude, they need to be in between "".}

\item{variable}{Give the variable that you want, either g,rh, or t. You can only give one of them and they need to be in between "".}

\item{folder}{Give the path of your folder containing the data for the atmospheric variables. Here you do need to give the whole path.}

\item{itx3d}{Give a previously generated array with the record series for all stations as binary indicators.}

\item{stations}{Give a previously generated data frame containing data for the stations (names, latitude, longitude) considered, they must be in the same order as in itx3d.}

\item{days}{Give a vector of two elements with the firs day of the year you want to get data for and the last day of the year you want to get data for. It has to be in the format "%m-%d" where the month goes first and the day of the month second, and they are separated by a - E.g. c("06-01", "08-31")}
}
\value{
A data frame
}
\description{
Get data for the atmospheric variable of your choice and the altitudes of your choice for the closest and the furthest four surrounding points of your station of choice
}
\examples{
atm_var_station("ZARAGOZA", "Data", c("300", "500", "700"), "g", "C:/Users/elsac/Documents/ERA5-data/Data/geo.data/", itx3d, stations, c("06-01", "08-31"))
}
