% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_atmospheric_data.R
\name{get_atmospheric_data}
\alias{get_atmospheric_data}
\title{Get an array of data for an atmospheric variable}
\usage{
get_atmospheric_data(folder_path, time, variable, geo_level, lat, lon, days)
}
\arguments{
\item{folder_path}{Give the path of the folder where the data files are kept.}

\item{time}{Give a specific time E.g."12pm" It has to be in between "" and have the 12-hour-clock format.}

\item{variable}{Specify the variable "rh" for relative humidity, "g" for geopotential, "t" for temperature.}

\item{geo_level}{Give a value of the pressure level without units, E.g "300".}

\item{lat}{Give a value for latitude E.g 40.}

\item{lon}{Give a value for longitude E.g. -10}

\item{days}{Give a vector of two elements with the firs day of the year you want to get data for and the last day of the year you want to get data for. It has to be in the format "%m-%d" where the month goes first and the day of the month second, and they are separated by a - E.g. c("06-01", "08-31").}
}
\value{
An array with the data for an atmospheric variable in a geopotential level for a specific time in a specific point and for a range of days of the year. Build the file name
}
\description{
Get an array of data for an atmospheric variable
}
\examples{
get_atmospheric_data("C:/Users/elsac/Documents/ERA5-data/Data/geo.data/", "12pm", "rh", "300", 40, -10, c("06-01", "08-31"))

}
